Changes in component, grid and state mainly pertain to the ordering of levels. Unsure if the other examples will work now!
setup.py has been changed to be able to build the dynamical core
using make. This will probably change in the future.

The dynamical core provides initial conditions, so currently
dynamics.py passes these on to Component during __init__

Due to some issue I faced with stepping the fields forward
in time using CliMT's leapfrog, all fields are being
stepped forward using the dynamical core and output as
diagnostic and not prognostic fields. This keeps the code
stable and does not blow up.